backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
ultralytics==8.3.0
easyocr==1.7.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
albumentations==1.4.15
torch>=2.0.0
torchvision>=0.15.0
python-dotenv==1.0.1
pydantic==2.8.2
